requirements.txt
faster-whisper>=1.0.0
sounddevice>=0.4.6
numpy>=1.24.0
keyboard>=0.13.5
pyautogui>=0.9.54
pyperclip>=1.8.2
pystray>=0.19.5
Pillow>=10.0.0
PyYAML>=6.0
requests>=2.31.0
winotify>=1.1.0
soundfile>=0.12.0
silero-vad>=5.1
onnxruntime>=1.18
scipy>=1.11

